Update PedDetection.md
Hi Matthew, 

Here is Boulder AI's feedback on the ATCMTD materials on Github.  On the whole, this is exactly the sort of system we could provide and we added a few comments to provide feedback on technology capabilities that the city should look for/expect.  Great job on the Github materials and please let me know if you have any questions.